Configure for 500*500 & Update gitignore

+++ data/patch_extractor.py
@@ -63,8 +63,8 @@ def make_grid(tile_size, patch_size, overlap):
     max_h = tile_size[0] - patch_size[0]
     max_w = tile_size[1] - patch_size[1]
     if max_h > 0 and max_w > 0:
-        h_step = np.ceil(tile_size[0] / (patch_size[0] - overlap))
-        w_step = np.ceil(tile_size[1] / (patch_size[1] - overlap))
+        h_step = int(np.ceil(tile_size[0] / (patch_size[0] - overlap)))
+        w_step = int(np.ceil(tile_size[1] / (patch_size[1] - overlap)))
     else:
         h_step = 1
         w_step = 1
